Do not install pulumi automatically
Pulumi is marked as a requirement, but I do not expect a Makefile to install it for me using `curl | bash`. I would prefer to get an error and figure out myself how to install Pulumi. Notably, if I have an environment with Pulumi but I forgot to switch to it, I don't expect to have another copy installed this way.